% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_trend.R
\name{form_beta_check}
\alias{form_beta_check}
\title{Beta Specification Check}
\usage{
form_beta_check(form, beta, IP, timepoints)
}
\arguments{
\item{form}{character value specifying the type of time trend. Options
include 'linear', 'quadratic', 'cubic', 'M', 'W', 'L_up', and 'L_down'.}

\item{beta}{vector specifying the appropriate parameters for functional
trend. See details of \code{\link{mean_trend}} for explanation for each form}

\item{IP}{vector specifying the inflection points. See details of
\code{\link{mean_trend}} for explanation for each form}

\item{timepoints}{numeric vector specifying the points to fit the functional
trend.

@keywords internal}
}
\value{
Nothing returned unless an error is returned.
}
\description{
Function for checking that the appopriate beta parameters are specified for
each of the mean trend specifications
}
