% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_norm_microbiome.R
\name{gen_microbiome_norm_feature_check}
\alias{gen_microbiome_norm_feature_check}
\title{Checking that features are specified appopriately}
\usage{
gen_microbiome_norm_feature_check(features, diff_abun_features)
}
\arguments{
\item{features}{Numeric value specifying the total number of features to
simulate in the microbiome. Must be greater than zero}

\item{diff_abun_features}{Number of features to simulate with differentially
abundant pattern. Must be between zero and number of features specified}
}
\value{
Potential warning message if no differentially abundant features or all
differentially abundant features are specified
}
\description{
Checking that features are specified appopriately
}
