% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_norm_microbiome.R
\name{gen_norm_microbiome}
\alias{gen_norm_microbiome}
\title{Generate Longitduinal Differential Abundance from Multivariate Normal}
\usage{
gen_norm_microbiome(features = 10, diff_abun_features = 5, n_control,
  n_treat, control_mean, sigma, num_timepoints, rho, corr_str = c("ar1",
  "compound", "ind"), func_form = c("linear", "quadratic", "cubic", "M",
  "W", "L_up", "L_down"), beta, IP = NULL, missing_pct,
  missing_per_subject, miss_val = NA, dis_plot = FALSE,
  plot_trend = FALSE, zero_trunc = TRUE)
}
\arguments{
\item{features}{numeric value specifying the number of features/microbes to
simulate. Default is 10.}

\item{diff_abun_features}{numeric value specifying the number of
differentially abundant features. Default is 5.}

\item{n_control}{integer value specifying the number of control individuals}

\item{n_treat}{integer value specifying the number of treated individuals}

\item{control_mean}{numeric value specifying the mean value for control
subjects. all control subjects are assummed to have the same population mean
value.}

\item{sigma}{numeric value specifying the global population standard
deviation for both control and treated individuals.}

\item{num_timepoints}{integer value specifying the number of timepoints per
subject.}

\item{rho}{value for the correlation parameter. must be between [0, 1].
see \code{\link[microbiomeDASim]{mvrnorm_corr_gen}} for details.}

\item{corr_str}{correlation structure selected. see
\code{\link[microbiomeDASim]{mvrnorm_corr_gen}} for details.}

\item{func_form}{character value specifying the functional form for the
longitduinal mean trend. see \code{\link[microbiomeDASim]{mean_trend}}
for details.}

\item{beta}{vector value specifying the parameters for the differential
abundance function. see \code{\link[microbiomeDASim]{mean_trend}} for
details.}

\item{IP}{vector specifying any inflection points. depends on the type of
functional form specified. see \code{\link[microbiomeDASim]{mean_trend}} for
details. by default this is set to NULL.}

\item{missing_pct}{numeric value that must be between [0, \1] that specifies
what percentage of the individuals will have missing values.}

\item{missing_per_subject}{integer value specifying how many observations per
subject should be dropped. note that we assume that all individuals must
have baseline value, meaning that the maximum number of
\code{missing_per_subject} is equal to \code{num_timepoints} - 1.}

\item{miss_val}{value used to induce missingness from the simulated data.
by default missing values are assummed to be NA but other common choices
include 0.}

\item{dis_plot}{logical argument on whether to plot the simulated data or
not. by default plotting is turned off.}

\item{plot_trend}{specifies whether to plot the true mean trend. see
\code{\link[microbiomeDASim]{mean_trend}} for details.}

\item{zero_trunc}{logical indicator designating whether simulated outcomes
should be zero truncated. default is set to TRUE}
}
\value{
This function returns a list with the following objects

\code{Y} The full simulated feature sample matrix where each row represent a
feature and each column a sample. Note that the differential and
non-differential bugs are marked by row.names
}
\description{
Generate Longitduinal Differential Abundance from Multivariate Normal
}
\examples{
gen_norm_microbiome(features = 5, diff_abun_features = 2,
               n_control = 10, n_treat = 10, control_mean = 8, sigma = 1,
               num_timepoints = 5, rho = 0.8, corr_str = "compound",
               func_form = "linear", beta =  c(0, 1), missing_pct = 0.3,
               missing_per_subject = 2)

}
