% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_trend.R
\name{IP_form_check}
\alias{IP_form_check}
\title{Inflection point check for \code{mean_trend}}
\usage{
IP_form_check(form, beta, IP, timepoints)
}
\arguments{
\item{form}{character value specifying the type of time trend. Options
include 'linear', 'quadratic', 'cubic', 'M', 'W', 'L_up', and 'L_down'.}

\item{beta}{vector specifying the appropriate parameters for functional
trend. See details of \code{\link{mean_trend}} for explanation for each form}

\item{IP}{vector specifying the inflection points. See details of
\code{\link{mean_trend}} for explanation for each form}

\item{timepoints}{numeric vector specifying the points to fit the functional
trend.}
}
\value{
Updated inflection point vector
}
\description{
Inflection point check for \code{mean_trend}
}
