% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_trend.R
\name{mean_trend_beta_vec}
\alias{mean_trend_beta_vec}
\title{Create beta vector for \code{mean_trend} for all functional forms}
\usage{
mean_trend_beta_vec(form, beta, IP, timepoints)
}
\arguments{
\item{form}{character value specifying the type of time trend. Options
include 'linear', 'quadratic', 'cubic', 'M', 'W', 'L_up', and 'L_down'.}

\item{beta}{vector specifying the appropriate parameters for functional
trend. See details of \code{\link{mean_trend}} for explanation for each form}

\item{IP}{vector specifying the inflection points. See details of
\code{\link{mean_trend}} for explanation for each form}

\item{timepoints}{numeric vector specifying the points to fit the functional
trend.

@keywords internal}
}
\value{
Vector with beta values used to create mean_tend
}
\description{
Create beta vector for \code{mean_trend} for all functional forms
}
