% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_trend.R
\name{mean_trend_design_mat}
\alias{mean_trend_design_mat}
\title{Create Design Matrix for \code{\link{mean_trend}} function}
\usage{
mean_trend_design_mat(form, beta, IP, timepoints)
}
\arguments{
\item{form}{character value specifying the type of time trend. Options
include 'linear', 'quadratic', 'cubic', 'M', 'W', 'L_up', and 'L_down'.}

\item{beta}{vector specifying the appropriate parameters for functional
trend. See details of \code{\link{mean_trend}} for explanation for each form}

\item{IP}{vector specifying the inflection points. See details of
\code{\link{mean_trend}} for explanation for each form}

\item{timepoints}{numeric vector specifying the points to fit the functional
trend.}
}
\value{
Numeric matrix with values that will be used to generate functional trends
}
\description{
By taking in the user specified parameters, we can return a design matrix
to use when creating the differential longitudinal abundance.
}
\keyword{internal}
