% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvrnorm_sim.R
\name{sigma_corr_function}
\alias{sigma_corr_function}
\title{Generating the longitudinal correlation matrix for repeated observations}
\usage{
sigma_corr_function(obs, sigma, corr_str, rho)
}
\arguments{
\item{obs}{number of repeated observations per individual}

\item{sigma}{the standard deviation parameter for the covariance matrix}

\item{corr_str}{the type of correlatin structure chosen. options currently
available include "ar1", "compound", and "ind"}

\item{rho}{the correlation coefficient for non-independent structures}
}
\value{
Return the covariance matrix V as a list
}
\description{
Generating the longitudinal correlation matrix for repeated observations
}
\keyword{internal}
