% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvrnorm_sim.R
\name{trunc_bugs}
\alias{trunc_bugs}
\title{Function for inducing truncation of outcome}
\usage{
trunc_bugs(Y, N, Mu, Sigma, zero_trunc)
}
\arguments{
\item{Y}{The original N x 1 vector of simulated multivariate outcomes}

\item{N}{Total number of observations equal to sum of repeated measurements
for all individuals}

\item{Mu}{N x 1 vector representing the mean values}

\item{Sigma}{N x N numeric matrix representing the covariance matrix for
the feature}

\item{zero_trunc}{Logical indicator whether to perform zero-truncation}
}
\value{
Potentially truncated outcome vector Y
}
\description{
Function for inducing truncation of outcome
}
