% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvrnorm_sim.R
\name{vector_scalar_check}
\alias{vector_scalar_check}
\title{Checking input for scalar or vector valued}
\usage{
vector_scalar_check(input, n, N)
}
\arguments{
\item{input}{a variable that will be supplied that we want to either check
the length or replicate to the specified length}

\item{n}{an integer value specifying the desired vector length if a scalar
is provided}

\item{N}{an integer value specifying the total number of observations across
individuals.}
}
\value{
return a vector that has same length as specified n
}
\description{
This function allows users to specify just a scalar value as a parameter in
the generating function, and here it is converted to a vector
of the proper length.
}
\details{
Note that \code{n} refers to the number of individuals in the
\code{\link[microbiomeDASim]{mvrnorm_corr_gen}}
}
\keyword{internal}
