% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvrnorm_sim.R
\name{mvrnorm_corr_gen}
\alias{mvrnorm_corr_gen}
\title{Generate Multivariate Random Normal Longitudinal Data}
\usage{
mvrnorm_corr_gen(n, obs, mu, sigma, rho, corr_str = c("ar1", "compound",
    "ind"), zero_trunc = TRUE)
}
\arguments{
\item{n}{integer scalar representing the total number of individuals}

\item{obs}{integer or vector specifying the number of observations per
indivdiual. If an integer then all indivdiuals are assummed to have the same
number of obsevations. If a vector, then the vector must have length equal
to \code{n} where each element specifies the number of observations for
the \eqn{i^{th}} individual.}

\item{mu}{integer or vector specifying the mean value for individuals.
If an integer then all indivdiuals are assummed to have the same mean.
If a vector, then the vector must have length equal to \code{n} where each
element specifies the mean for the \eqn{i^{th}} individual.}

\item{sigma}{numeric scalar or vector specifying the standard deviation for
observations.}

\item{rho}{numeric scalar value between [0, 1] specifying the amount of
correlation between. assumes that the correlation is consistent for all
subjects.}

\item{corr_str}{character value specifying the correlation structure.
Currently available methods are 'ar1', 'compound', and 'ind' which
correspond to first-order autoregressive, compound or equicorrelation,
and independence respecitvely.}

\item{zero_trunc}{logical value to specifying whether the generating
distribution should come from a multivariate zero truncated normal or an
untruncated multivariate normal. by default we assume that zero truncation
occurs since this is assummed in our microbiome setting.}
}
\value{
This function returns a list with the following objects:

\code{df} - data.frame object with complete outcome \code{Y}, subject ID,
time, group, and outcome with missing data

\code{Y} - vector of complete outcome

\code{Mu} - vector of complete mean specifications used during simulation

\code{Sigma} - block diagonal symmetric matrix of complete data used during
simulation

\code{N} - total number of observations
}
\description{
For this methodology we assume that we draw a set of \code{n} independent each
with \eqn{q_{i}} observations.
}
\examples{
mvrnorm_corr_gen(n=15, obs=4, mu=20, sigma=2, rho=0.9, corr_str="ar1")

}
