% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_functions.R
\name{simulate2MRexperiment}
\alias{simulate2MRexperiment}
\title{Convert simulated output to MRexperiment object}
\usage{
simulate2MRexperiment(obj)
}
\arguments{
\item{obj}{}
}
\value{
An MRexperiment object
}
\description{
In order to allow investigators to more easily incorporate simulated data,
this package converts the raw output into an MRexperiment object used in the
\code{\link{metagenomeSeq}} package.
}
\examples{
gen_norm_microbiome()


}
